backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
PyPDF2
python-docx
python-pptx